Add wishbone classic arbiter.
The conversion from wishbone classic to wishbone pipelined to make use
of the ZipCPU arbiter adds a bunch of extra clock delays that strangle
the VexRiscV core from being able to fetch instructions fast enough.

We attempt to work around it with a new arbiter that is natively using
wishbone classic. This new arbiter delays the CYC signal by one clock
for arbitration, and the rest of the signals are then muxed using
combinatorial logic to try and eliminate any delays, and allows the
core to fetch an instruction every other cycle.